refactor: SheetCanvas 순수 canvas 유틸을 lib/canvas.ts로 추출
